refactor(bridge): clean up data types
