+++ bridge/src/ethereum-burn-event-monitor.ts
@@ -1,7 +1,7 @@
 import Web3 from "web3";
 import { Contract, EventData } from 'web3-eth-contract';
 import { ConfirmationMonitor } from "./confirmation-monitor";
-import { ContractDescription } from "./interfaces/contract-description";
+import { ContractDescription } from "./types/contract-description";
 import { TransactionLocation } from "./types/transaction-location";
 
 const BURN_EVENT_NAME = "Burn";
