+++ bridge/src/interfaces/headless-graphql-client.ts
@@ -1,12 +1,12 @@
 import { BlockHash } from "../types/block-hash";
 import { TxId } from "../types/txid";
-import { INCGTransferredEvent } from "./ncg-transferred-event";
+import { NCGTransferredEvent } from "../types/ncg-transferred-event";
 
 export interface IHeadlessGraphQLClient {
     getBlockIndex(blockHash: BlockHash): Promise<number>;
     getTipIndex(): Promise<number>;
     getBlockHash(index: number): Promise<BlockHash>;
-    getNCGTransferredEvents(blockHash: BlockHash, recipient: string): Promise<INCGTransferredEvent[]>;
+    getNCGTransferredEvents(blockHash: BlockHash, recipient: string): Promise<NCGTransferredEvent[]>;
     getNextTxNonce(address: string): Promise<number>;
     transfer(recipient: string, amount: string, txNonce: number, memo: string | null): Promise<TxId>;
 }
