+++ bridge/src/types/burn-event-result.ts
